Move usage of ResourceMap type to new spot
