+++ src/core/resources/resource.loader.ts
@@ -2,11 +2,11 @@ import { Injectable, Type } from '@nestjs/common';
 import { ConditionalKeys, ValueOf } from 'type-fest';
 import { ID, Many, ObjectView, ServerException } from '~/common';
 import { GqlContextHost } from '~/core/graphql';
-import { ResourceMap } from '../../components/authorization/model/resource-map';
 import { LoaderContextType, LoaderOf, NestDataLoader } from '../data-loader';
 import { NEST_LOADER_CONTEXT_KEY } from '../data-loader/constants';
 import { BaseNode } from '../database/results';
 import { ResourceLoaderRegistry } from './loader.registry';
+import { ResourceMap } from './map';
 import { ResourceResolver } from './resource-resolver.service';
 
 type SomeResourceType = ValueOf<ResourceMap>;
